Bugfix/Refactor: Better handling of subproc output
If the output streams for stdout or stderr get full, as defined by the
OS, the process can hang.  Waiting to read them until a command is run
is not enough.  The problem seems much more prevelant on Windows, where
this issue was discovered via the Simple R extension.  The solution is
to just pump the output as soon as we can via threads.